+++ shardingsphere-features/shardingsphere-encrypt/shardingsphere-encrypt-spring/shardingsphere-encrypt-spring-namespace/src/main/resources/META-INF/spring.handlers
@@ -15,7 +15,4 @@
 # limitations under the License.
 #
 
-http\://shardingsphere.apache.org/schema/shardingsphere/datasource=org.apache.shardingsphere.driver.spring.namespace.handler.DataSourceNamespaceHandler
-http\://shardingsphere.apache.org/schema/shardingsphere/sharding=org.apache.shardingsphere.driver.spring.namespace.handler.ShardingNamespaceHandler
-http\://shardingsphere.apache.org/schema/shardingsphere/masterslave=org.apache.shardingsphere.driver.spring.namespace.handler.MasterSlaveNamespaceHandler
-http\://shardingsphere.apache.org/schema/shardingsphere/spi=org.apache.shardingsphere.driver.spring.namespace.handler.SPINamespaceHandler
+http\://shardingsphere.apache.org/schema/shardingsphere/encrypt=org.apache.shardingsphere.encrypt.spring.namespace.handler.EncryptRuleNamespaceHandler
