+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-orchestration-spring/shardingsphere-jdbc-orchestration-spring-namespace/src/main/java/org/apache/shardingsphere/driver/orchestration/spring/namespace/constants/InstanceBeanDefinitionParserTag.java
@@ -34,7 +34,7 @@ public final class InstanceBeanDefinitionParserTag {
     
     public static final String SERVER_LISTS_TAG = "server-lists";
     
-    public static final String NAMESPACE_TAG = "namespace";
+    public static final String NAMESPACE_TAG = "META-INF/namespace";
     
     public static final String PROPERTY_REF_TAG = "props-ref";
 }
